created a store for fetching from the backend
